Move ingredient scaling functions from recipe2vec to utils
