+++ beer.ai/recipe2vec.py
@@ -7,6 +7,7 @@ import numpy as np
 import pandas as pd
 import pickle
 from tqdm import tqdm
+from .utils import scale_ferm, scale_hop, scale_misc, scale_yeast 
 
 CORE_COLS = ["batch_size", "boil_size", "boil_time", "efficiency"]
 ING_COLS = [
@@ -81,74 +82,6 @@ def recipes2vec(recipes):
     ]
 
     return recipes_vec
-
-
-def scale_ferm(df):
-    """
-    (DataFrame) -> float
-    Compute the scaled fermentable quantity.
-
-    Take as input a subset of the ing DataFrame, joined to the core DataFrame.
-    Replace ferm_scaled with the gravity contribution of the fermentable:
-        g/L extract in the boil kettle.
-    """
-    df["ferm_amount"] = (
-        df["ferm_amount"] * df["ferm_yield"] * df["efficiency"] / df["boil_size"]
-    )
-    df["ferm_amount"] = df["ferm_amount"].replace([np.inf, -np.inf], np.nan)
-    return df
-
-
-def scale_hop(df):
-    """
-    (DataFrame) -> float
-    Compute the scaled hop quantity.
-
-    Take as input a subset of the ing DataFrame, joined to the core DataFrame.
-    Return a different quantity depending on the use:
-        Dry hops:  dry hopping rate
-            grams of dry hops per litre in the batch
-        Boil hops: AUU
-            grams of alpha acids per litre in the boil kettle
-    """
-    # Dry hops
-    dh_cond = df["hop_use"] == "dry hop"
-    df.loc[dh_cond, "hop_amount"] = (
-        df.loc[dh_cond, "hop_amount"] / df.loc[dh_cond, "batch_size"]
-    )
-
-    # Every other hop use
-    bh_cond = df["hop_use"] != "dry hop"
-    df.loc[bh_cond, "hop_amount"] = (
-        df.loc[bh_cond, "hop_amount"]
-        * df.loc[bh_cond, "hop_alpha"]
-        * (1 - 0.1 * (df.loc[bh_cond, "hop_form"] == "leaf").astype(int))
-        / df.loc[bh_cond, "boil_size"]
-    )
-    df["hop_amount"] = df["hop_amount"].replace([np.inf, -np.inf], np.nan)
-    return df
-
-
-def scale_misc(df):
-    """
-    (DataFrame) -> float
-    Compute the scaled misc quantity.
-
-    Take as input a subset of the ing DataFrame, joined to the core DataFrame.
-    Return a scaled misc quantity.
-    """
-    df["misc_amount"] = df["misc_amount"] / df["batch_size"]
-    df["misc_amount"] = df["misc_amount"].replace([np.inf, -np.inf], np.nan)
-    return df
-
-
-def scale_yeast(df):
-    """
-    (DataFrame) -> DataFrame
-    Return a DataFrame with a new column yeast_amount = 1
-    """
-    df.loc[~df["yeast_name"].isna(), "yeast_amount"] = 1
-    return df
 
 
 def scale_quantities(df):
